moving split and join_three into string_util.c and adding a headerfile
